+++ src/main/java/com/stellar/coreBasic/entity/HighSchoolGrade.java
@@ -1,4 +1,4 @@
-package com.stellar.coreBasic.member;
+package com.stellar.coreBasic.entity;
 
 public enum HighSchoolGrade {
     FRESHMAN,JUNIOR,SENIOR
