+++ src/main/java/com/stellar/coreBasic/repository/MemoryMemberRepository.java
@@ -1,10 +1,11 @@
-package com.stellar.coreBasic.member;
+package com.stellar.coreBasic.repository;
 
-import java.util.HashMap;
+import com.stellar.coreBasic.entity.Member;
+
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
-public class MemoryMemberRepository implements MemberRepository{
+public class MemoryMemberRepository implements MemberRepository {
     private static Map<String, Member> store = new ConcurrentHashMap<>();
     @Override
     public void save(Member member) {
